+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ## [0.7.0]
 ### Added
+- [#36](https://github.com/picimako/mockitools/issues/36): Added inspection to report `toString()` calls in verifications, since Mockito cannot verify `toString()`.
 - [#39](https://github.com/picimako/mockitools/issues/39): Added inspection to report `*Return()` stubbing calls, if the stubbed method's return type is void.
 - [#40](https://github.com/picimako/mockitools/issues/40): Added inspection to report `doNothing()` and `willDoNothing()` stubbing calls,
   if the stubbed method's return type is not void.
@@ -13,6 +14,7 @@
 ### Changed
 - [#41](https://github.com/picimako/mockitools/issues/41): Cleaned up the class and package name, and modified the icons in the target method selection
   popup during mock field conversion. The new icons can help distinguish between before hooks, test methods, and the rest of the methods.
+- Refactorings and housekeeping in the mock stubbing and verification areas.
 
 ## [0.6.0]
 ### Added